Add the images, create main function, now all instances work except niki cat

+++ projet/phase5/tools.jl
@@ -2,8 +2,8 @@ using Random, FileIO, Images, ImageView, ImageMagick
 
 """Compute the similarity score between two pixels."""
 function compare_pixels(p1, p2)
-	r1, g1, b1 = red(p1), green(p1), blue(p1)
-	r2, g2, b2 = red(p2), green(p2), blue(p2)
+	r1, g1, b1 = Float64(red(p1)), Float64(green(p1)), Float64(blue(p1))
+	r2, g2, b2 = Float64(red(p2)), Float64(green(p2)), Float64(blue(p2))
 	return abs(r1-r2) + abs(g1-g2) + abs(b1-b2)
 end
